test(scripts): enforce the pure LOC ceiling on every script file
Split scripts/runtime-public-api-snapshot.mjs into runtime-public-api-collect.mjs and runtime-public-api-snapshot.mjs so both modules stay below the 250 pure-LOC ceiling.

Ultraworked with [omo](https://github.com/code-yeongyu/oh-my-openagent)
